--- src/main/java/uv/fei/bussinesslogic/IExpertoDAO.java
@@ -1,10 +1,0 @@
-package uv.fei.bussinesslogic;
-
-import uv.fei.domain.Experto;
-
-import java.sql.SQLException;
-
-public interface IExpertoDAO {
-    public boolean agregarExperto(Experto experto) throws SQLException;
-
-}
