+++ src/main/java/uv/fei/domain/Usuario.java
@@ -2,8 +2,36 @@ package uv.fei.domain;
 
 public class Usuario {
     private int id;
+    private String acreditacion;
+    private String rol;
+    private String email;
     private String nombre;
     private String contrasenia;
+
+
+    public String getAcreditacion() {
+        return acreditacion;
+    }
+
+    public void setAcreditacion(String acreditacion) {
+        this.acreditacion = acreditacion;
+    }
+
+    public String getRol() {
+        return rol;
+    }
+
+    public void setRol(String rol) {
+        this.rol = rol;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
 
     public int getId() {
         return id;
@@ -31,12 +59,11 @@ public class Usuario {
 
     public Usuario() {
         this.id = 0;
+        this.acreditacion = "";
+        this.rol = "";
+        this.email = "";
         this.nombre = "";
         this.contrasenia = "";
     }
 
-    public Usuario(String nombre, String contrasenia) {
-        this.nombre = nombre;
-        this.contrasenia = contrasenia;
-    }
 }
